FROM_UPSTREAM: iio: hid-sensor-hub: Remove hard coded indexes
Remove the hard coded indexes, instead search for usage id and
use the index to set the power and report state.
This will fix issue, where the report descriptor doesn't contain
the full list of possible selector for power and report state.

Adding the hid-sensor-hub.h change which was missed in the merge
of https://android.intel.com/#/c/199928.

Upstream-commit-id: 1a214ae5d1a66ead18c5e8c3a8d03b969b0ee647

Change-Id: I1adc2e9e4613f5fa32d25ff25e1b0f52e6c8dd5c
Issue: AREQ-297
Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Acked-by: Jonathan Cameron <jic23@kernel.org>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>