REVERTME [VPG]: drm/i915: Do not use turbo in BDW C stepping.
Enabling turbo leads to GPU hangs in BDW C stepping.

The problem is that  MCHBAR_PCU_C0 is not recorded in BDW C steppings
(that’s MCHBAR_MIRROR_BASE + 0x5960, the number of cycles that GT was in
C0).

Because of this problem, current_c0 is always zero and
bdw_sw_calculate_freq() won’t work. The ucore register is OK in D
stepping and turbo operates correctly there.

Tagged as REVERTME as disabling turbo affects performance considerably,
but this workaround is needed until C steppings are phased out.

Signed-off-by: Michel Thierry <michel.thierry@intel.com>

Pulled from ABT/kernel-internal to GMin/gmin.

Change-Id: Ie9a4cd284f0df5b38da1097e3f43679da8ee0539
Signed-off-by: Oscar Mateo <oscar.mateo@intel.com>