BACKPORT: mmc: Allow forward compatibility for eMMC
As stated by the eMMC 5.0 specification, a chip should not be rejected
only because of the revision stated in the EXT_CSD_REV field of the
EXT_CSD register.

Remove the control on this value, the control of the CSD_STRUCTURE field
should be sufficient to reject future incompatible changes.

Signed-off-by: Romain Izard <romain.izard.pro@gmail.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>

Change-Id: I7a0ab1a38b443a68e6c70b47a628a82a3dc66c75
Tracked-on: https://jira01.devtools.intel.com/browse/IMINAN-8269
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>